requirements.txt
absl-py==1.0.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
ale-py==0.7.5
annotated-types==0.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blessed==1.20.0
box2d-py==2.3.8
cachetools==5.0.0
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chex==0.1.3
click==8.1.3
cloudpickle==2.0.0
cmake==3.26.4
colorama==0.4.6
colorful==0.5.5
ConfigSpace==0.6.1
cycler==0.11.0
dask==2023.3.0
dask-jobqueue==0.8.2
dataclasses==0.6
debugpy==1.5.1
decorator==4.4.2
deepdiff==5.8.1
defusedxml==0.7.1
distlib==0.3.6
distrax==0.1.2
distributed==2023.3.0
dm-acme==0.2.4
dm-env==1.5
dm-haiku==0.0.6
dm-launchpad==0.3.2
dm-reverb==0.6.1
dm-sonnet==2.0.0
dm-tree==0.1.7
dnspython==2.2.1
docopt==0.6.2
emcee==3.1.4
entrypoints==0.4
enum34==1.1.10
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.12.2
flatbuffers==2.0.7
flax==0.5.0
fonttools==4.33.3
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gpustat==1.1
grpcio==1.51.3
gviz-api==1.10.0
gym==0.21.0
gym-notices==0.0.6
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h5py==3.6.0
HeapDict==1.0.1
id-mava @ git+https://github.com/instadeepai/Mava@4b07ae57f90f1207faef8957241767331a8746d8
idna==3.3
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==8.0.6
jax==0.3.10
jaxlib==0.3.10
jedi==0.18.1
Jinja2==3.1.2
jmp==0.0.2
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.5.1
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaitaistruct==0.10
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy_loader==0.3
libclang==14.0.1
lit==16.0.6
locket==1.0.0
lz4==4.3.2
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
memory-profiler==0.61.0
mistune==0.8.4
mlagents==0.28.0
mlagents-envs==0.28.0
mock==4.0.3
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
mpyq==0.2.5
msgpack==1.0.4
multi-agent-ale-py==0.1.11
multidict==6.0.4
munch==2.5.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
networkx==3.1
notebook==6.4.11
numpy==1.21.4
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
optax==0.1.2
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
PettingZoo==1.18.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pkg_resources==0.0.0
platformdirs==3.8.1
portpicker==1.5.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.3.0
pydot==1.4.2
pygame==2.1.0
Pygments==2.15.1
pymongo==4.3.2
pymunk==6.2.1
pynisher==1.0.5
pyparsing==3.0.8
pyrfr==0.9.0
pyrsistent==0.18.1
PySC2==3.0.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==22.3.0
ray==2.6.1
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.4.2
rlax==0.1.2
rsa==4.8
ruamel.base==1.0.0
ruamel.std.pathlib==0.9.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
ryd==0.7.1
s2clientprotocol==5.0.9.87702.0
s2protocol==5.0.9.87702.0
sacred==0.8.4
scikit-image==0.21.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sk-video==1.1.10
smart-open==6.3.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.12
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-profile==2.11.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.1
tensordict==0.0.3
tensorflow==2.7.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-probability==0.15.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.1.1
tomli==2.0.1
toolz==0.11.2
torch==1.13.1
torchrl==0.0.5
torchsummary==1.5.1
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
trfl==1.2.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
urllib3==1.26.9
virtualenv==20.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
whichcraft==0.6.1
widgetsnbextension==4.0.7
wrapt==1.14.1
yarl==1.9.2
zict==2.2.0
zipp==3.8.0
